+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiRenderer.kt
@@ -1,0 +1,15 @@
+package org.simple.clinic.home.overdue.search
+
+import org.simple.clinic.mobius.ViewRenderer
+
+class OverdueSearchUiRenderer(
+    private val ui: OverdueSearchUi
+) : ViewRenderer<OverdueSearchModel> {
+
+  override fun render(model: OverdueSearchModel) {
+    if (!model.hasSearchQuery) {
+      ui.showSearchHistory(model.overdueSearchHistory.orEmpty())
+      ui.hideSearchResults()
+    }
+  }
+}