+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Bump Annotation library to v1.4.0
 - Provide overdue search history string set from preferences
 - Add effect to load overdue search history
+- Show overdue search history when search query is empty
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291